Add event for refreshing the grains only
There are use cases when we want to refresh only the grains
without refreshing the other modules. One example would be when
locking a system. We want to lock the system, then wait for grains
to get refreshed in order to prevent the execution of the next jobs
and avoid race conditions.

Signed-off-by: Cristian Hotea <cristian.hotea@ni.com>